Pitch/Roll, Bosch Sensors
Added Pitch/Roll and new Classes for the Bosch sensors

+++ OpenDrone-flight-controller/OpenDrone_FC/Sensor/HMC5883L.cpp
@@ -31,9 +31,9 @@ short HMC5883L::readRawData(int addr)
 	return value;
 }
 
-int *HMC5883L::getMagnetometerValues()
+double *HMC5883L::getMagnetometerValues()
 {
-	static int ar[3];  /* Declared locally, but saved in the data-segment (and not on the stack) */
+	static double ar[3];  /* Declared locally, but saved in the data-segment (and not on the stack) */
 	ar[0] = readRawData(0x03); //Magnet X
 	ar[1] = readRawData(0x05); //Magnet Y
 	ar[2] = readRawData(0x07); //Magnet Z
